added room layout, control elements, notes and item inspector
Added a new customizable room layout as tilemap (textures can be replaced).

The current scene can be played as a 'game'. To get the first hint, click on the note (white paper on the upper side of the window).

Added new control elements such as a drawer, in which you can find 1 item (note, key,...), a table, which is currently useless and an exit door, which can be opened with a key.
These Elements dont have a texture yet.

Also added an item inspector. If an item is clicked this inspector opens as a window and shows a description of the current item or a text.

Also organized the hierarchy view